also done for send activity
