+++ src/cmd/anilist_module/anime_activity/send_activity.rs
@@ -1,7 +1,4 @@
-use std::collections::HashMap;
 use std::env;
-use std::fs::File;
-use std::io::Read;
 use std::thread::sleep;
 use std::time::Duration;
 
@@ -9,12 +6,12 @@ use chrono::Utc;
 use serenity::http::Http;
 use serenity::model::channel::Embed;
 use serenity::model::prelude::Webhook;
+use serenity::utils::Colour;
 use sqlx::FromRow;
 
 use crate::cmd::anilist_module::anime_activity::struct_minimal_anime::MinimalAnimeWrapper;
-use crate::cmd::general_module::get_guild_langage::get_guild_langage;
-use crate::cmd::general_module::lang_struct::SendActivityLocalisedText;
 use crate::cmd::general_module::pool::get_pool;
+use crate::cmd::lang_struct::embed::anilist::anilist_activity::struct_lang_send_activity::SendActivityLocalisedText;
 
 #[derive(Debug, FromRow, Clone)]
 pub struct ActivityData {
@@ -67,28 +64,14 @@ pub async fn send_activity() {
     for row in rows {
         if Utc::now().timestamp().to_string() != row.timestamp.clone().unwrap() {} else {
             let row2 = row.clone();
-            let mut file = File::open("lang_file/embed/anilist/anime_activity/send_activity.json")
-                .expect("Failed to open file");
-            let mut json = String::new();
-            file.read_to_string(&mut json).expect("Failed to read file");
-
-            let json_data: HashMap<String, SendActivityLocalisedText> =
-                serde_json::from_str(&json).expect("Failed to parse JSON");
-
-            let guild_id = row.server_id.clone().unwrap();
-            let lang_choice = get_guild_langage(guild_id.clone()).await;
-
+                let guild_id = row.server_id.clone();
             if row.delays.unwrap() != 0 {
                 tokio::spawn(async move {
-                    if let Some(localised_text) = json_data.get(lang_choice.as_str()) {
                         sleep(Duration::from_secs((row2.delays.clone().unwrap()) as u64));
-                        send_specific_activity(row, localised_text.clone(), guild_id, row2).await
-                    }
+                        send_specific_activity(row, guild_id.unwrap(), row2).await
                 });
             } else {
-                if let Some(localised_text) = json_data.get(lang_choice.as_str()) {
-                    send_specific_activity(row, localised_text.clone(), guild_id, row2).await
-                }
+                    send_specific_activity(row, guild_id.unwrap(), row2).await
             }
         }
     }
@@ -118,10 +101,11 @@ pub async fn update_info(row: ActivityData, guild_id: String) {
 
 pub async fn send_specific_activity(
     row: ActivityData,
-    localised_text: SendActivityLocalisedText,
     guild_id: String,
-    row2: ActivityData,
+    row2: ActivityData
 ) {
+    let color = Colour::FABLED_PINK;
+    let localised_text = SendActivityLocalisedText::get_add_activity_localised(guild_id.clone()).await.unwrap();
     let my_path = "./.env";
     let path = std::path::Path::new(my_path);
     let _ = dotenv::from_path(path);
@@ -144,6 +128,7 @@ pub async fn send_specific_activity(
                 row.name.unwrap(),
                 localised_text.end,
             ))
+            .color(color)
     });
     webhook
         .execute(&http, false, |w| w.embeds(vec![embed]))
