+++ src/cmd/lang_struct/embed/anilist/anilist_activity/mod.rs
@@ -1,1 +1,2 @@
 pub mod struct_lang_add_activity;
+pub mod struct_lang_send_activity;
